Blacklist update.
The last version had a serious blacklist bug in it so don't use it. The one is the first to attempt blacklisting the bad ones and get past simply not failing.

I also extended the initial blacklist time from 10 secs to 15, this allows a lil more headroom, it will also make it a lil more resilient when you have a botched behavior.