Added support for resurrection! You must be healing enabled and it must be in your Patrol routine.
Ninja Skin is also in... just waiting on button.  I made the var just like the skinningCheckbox... I believe I named it ninjaSkinCheckbox or somethin like that. It's already implemented in the BotController :)

Also fixed the revived bug... should get right to it when it revives.  Also... instead of going directly into post combat it evaluates incase you revived into a bad situation.