Blacklisting is ironed out now and working well.  There was a condition that would cause it to crash due to a bad loop/condition.
I added support for the detection of 'a more powerful spell is active'.  It doesn't perm blacklist the target, but it puts an error in the log file n tells the user to adjust their behavior.

With respect to RedWines log type comment.  I've ruffed in the new logging system, my objective is not to fully refine the log levels.  I've done a massive amount of work already... log levels should be gone through after I'm done rearranging.  I have adjusted it so that the 'non extended' version looks pretty, but aside from that it's only half way done (although you will not that I've made hundreds of changes to the log calls in many files).  

A handy suggestion... I think it would help to have log checkboxes for Blacklisting and Evaluation.  When I redid the logging method I made it inherent and I also adjusted the log level names to match your buttons.  If you follow your current naming convention and double check the LogController.h to confirm the name you should be all set... I don't think you'll have to add any code aside from the button :)