This started as a rewrite to PvP with the intention of converting it to an evaluation, but the scope has gone further than that in the persuit of debugging.
It's not perfect yet, but you can see where i'm going with this.

Evaluation will now halt it's self if called when performing a procedure, this has cleaned up a lot (this also means all procedures must restart evaluation).  Movement now needs some work because it's not being pushed from the BotController.